requirements.txt
python-docx
python-pptx
PyPDF2
pytesseract
pdf2image
Pillow